DELIA-40553: Duplicate timedMetadata reported and wrong time value
Reason for change: During linear playlist refresh , already reported Metadata
was getting reported .Also intermittenly time of Metadata was showing outside
culling window. Fixed the duplicate handling across playlist refresh.On underflow
retune, culled value was resetted to 0 causing timed Metadata time to reset.
Test Procedure: Tested with Peacock linear content and checked Metadata
Risks: Low

(cherry picked from commit d147f3cfb18520f68d5576d22ca0823257d56193)

+++ priv_aamp.h
@@ -1376,10 +1376,10 @@ public:
 	 * @param[in] name - Metadata name
 	 * @param[in] content - Metadata content
 	 */
-	TimedMetadata(double timeMS, std::string name, std::string content, std::string id, double durMS) : _timeMS(timeMS), _name(name), _content(content), _id(id), _durationMS(durMS) {}
+	TimedMetadata(long long timeMS, std::string name, std::string content, std::string id, double durMS) : _timeMS(timeMS), _name(name), _content(content), _id(id), _durationMS(durMS) {}
 
 public:
-	double      _timeMS;     /**< Time in milliseconds */
+	long long _timeMS;     /**< Time in milliseconds */
 	std::string _name;       /**< Metadata name */
 	std::string _content;    /**< Metadata content */
 	std::string _id;         /**< Id of the timedMetadata. If not available an Id will bre created */
@@ -2069,7 +2069,7 @@ public:
 	 * @param[in] durationMS - Duration in milliseconds
 	 * @return void
 	 */
-	void ReportTimedMetadata(double timeMS, const char* szName, const char* szContent, int nb, const char* id = "", double durationMS = -1);
+	void ReportTimedMetadata(long long timeMS, const char* szName, const char* szContent, int nb, const char* id = "", double durationMS = -1);
 
 	/**
 	 * @brief sleep only if aamp downloads are enabled.
